Clean Up Errors; Bootstrap Outside of Shared Dir
Removed all red text / error messages from the provisioning.

Now bootstrapping in a VM-local directory, rather than in a shared directory. This is higher performance and should avoid errors on Windows.

Update the README to remove the Windows Administrator note and to provide a one-liner for creating the machine and running the first build.