GUI Machine + Working Ada Build
Added CONTRIBUTING.md

Created a patch for GNAT community that updates the formal containers so that the Ada build succeeds.

Updated the Vagrantfile to further DRY up the provisioning. Then added a GUI machine to both uxas and uxas-almost. This machine will create an Ubunut desktop environment.

The README has been updated to describe how to use it.