Update Vagrantfile, README; new run-example.py
Created `run-example.py`, which is a frontend for running the examples packaged with OpenUxAS. Importantly, it allows paths to OpenUxAS and OpenAMASE to be specified on the command line or in environment variables; it defaults to using the correct location inside the anod sandbox. This has been tested reasonably well.

Updated the Vagrantfile to remove the -almost machines; instead, the Vagrantfile looks for the gnat-bin binary and maps it if it's present on the host. Then, during provisioning, it checks for the file before downloading it. This eliminates potential for errors from replicating the specs (especially since I only test the -almost variants).

Updated the GUI machine to set clipboard sharing; this works (the menu is set) but doesn't work (clipboard sharing is broken).

Finally, added a one-shot script that builds the GUI machine and runs all the builds.