Fixes #5 and #6: update README & build-vagrant-gui
Fix for issue #5 (Remove Reference to `vagrant-vbguest` Plugin): update the README to remove this optional step and update the the script to reload the gui machine after provisioning, so that the manually-installed guest additions will be reloaded.

Fix for issue #6 (README Should Warn about VirtualBox 6.1 Incompatibility): update the README to clarify instructions for downloading VirtualBox, including updating the address from which VirtualBox should be downloaded.