When creating the python venv, clear first
This fixes a potential (and observed) problem that arises if the user has a vpython directory in their host's copy of OpenUxAS-bootstrap. In this case, the vpython gets copied to the guest's bootstrap directory and is in the way of the venv creation. Unfortunately, when this happens, the error messages are extremely unhelpful and misleading.

Clearing the venv seems like a reasonable thing to do.