Feature/improve setenv (#37)
* Expose more paths through anod-setenv

Update the specs for:

- amase
- uxas
- uxas-ada

so that anod-setenv exposes more details of where things are in the sandbox. These changes should make it easier for users to find where anod is putting things. These changes also support updates to the run-examples script, which will migrate to OpenUxAS.

* Add needed dependency for running uxas-ada

The uxas-ada binary is not statically linked, so the library for zeromq needs to be on the library path.