Better output during provisioning
Update the Vagrantfile so that during provisioning, the user sees [Success] or [Error] and the name of the last command. The command itself will write all output to a log file in the VM. This way, if something goes wrong, the user can log into the VM and get the full log, rather than having to hunt through and capture data from the console output.

This is probably an intermediate state along the path of having a python-based provisioning script that is invoked after the VM comes up.