Provide autocompletion for run-example (#64)
Users will have to load this into their shell by sourcing the file,
either directly (at the command line) or in their .bashrc, like this:

	source /path/to/bootstrap/install/run-example-completion.bash

Then tab completion will show the output of --list in run-example.

Note that --list isn't actually entirely well-suited for this, so there
is extra garbage that is offered. A change would be needed to
run-example to give it a mode specifically for completion. This does
work, at present.

This also works for zsh.