Create anod command to generate VS Code config
This commit adds a new command to anod: `anod configure-vscode`. When
run, this command will create the file `c_cpp_properties.json` and place
it, by default, in `develop/OpenUxAS/.vscode`. The configuration will
include the necessary include search paths for the spec given.

The envisioned use case is during OpenUxAS development. When `anod
devel-setup` is run, `anod configure-vscode` is automatically run: this
yields the correct c_cpp_properties.json to provide full intellisense
for OpenUxAS development.

If desired, the user can override the default destination or can output
the configuration to STDOUT.